update to 6.2.0

+++ src/app/code/community/Ess/M2ePro/Model/Connector/Ebay/Item/Stop/Multiple.php
@@ -49,7 +49,38 @@ class Ess_M2ePro_Model_Connector_Ebay_Item_Stop_Multiple
 
                     $this->getLogger()->logListingProductMessage($listingProduct, $message,
                                                                  Ess_M2ePro_Model_Log_Abstract::PRIORITY_MEDIUM);
+                } elseif ($listingProduct->isLockedObject(NULL) ||
+                          $listingProduct->isLockedObject('in_action')) {
+
+                    $message = array(
+                        // M2ePro_TRANSLATIONS
+                        // Another action is being processed. Try again when the action is completed.
+                        parent::MESSAGE_TEXT_KEY => 'Another action is being processed. '
+                                                   .'Try again when the action is completed.',
+                        parent::MESSAGE_TYPE_KEY => parent::MESSAGE_TYPE_ERROR
+                    );
+
+                    $this->getLogger()->logListingProductMessage($listingProduct, $message,
+                                                                 Ess_M2ePro_Model_Log_Abstract::PRIORITY_MEDIUM);
                 }
+
+                $this->failedListingProductIds[] = $listingProduct->getId();
+                continue;
+            }
+
+            if ($listingProduct->isLockedObject(NULL) ||
+                $listingProduct->isLockedObject('in_action')) {
+
+                $message = array(
+                    // M2ePro_TRANSLATIONS
+                    // Another action is being processed. Try again when the action is completed.
+                    parent::MESSAGE_TEXT_KEY => 'Another action is being processed. '
+                                               .'Try again when the action is completed.',
+                    parent::MESSAGE_TYPE_KEY => parent::MESSAGE_TYPE_ERROR
+                );
+
+                $this->getLogger()->logListingProductMessage($listingProduct, $message,
+                                                             Ess_M2ePro_Model_Log_Abstract::PRIORITY_MEDIUM);
 
                 $this->failedListingProductIds[] = $listingProduct->getId();
                 continue;
@@ -61,6 +92,12 @@ class Ess_M2ePro_Model_Connector_Ebay_Item_Stop_Multiple
             if (isset($this->params['remove']) && (bool)$this->params['remove']) {
 
                 foreach ($this->listingsProducts as $listingProduct) {
+
+                    if ($listingProduct->isLockedObject(NULL) ||
+                        $listingProduct->isLockedObject('in_action')) {
+                        continue;
+                    }
+
                     /** @var $listingProduct Ess_M2ePro_Model_Listing_Product */
                     $listingProduct->addData(array('status'=>Ess_M2ePro_Model_Listing_Product::STATUS_STOPPED))->save();
                     $listingProduct->deleteInstance();
